refactor(agent-wire) P1: relocate shared domain types out of modes/rpc/rpc-types
Consolidation plan phase 1 (no deletion). Move shared-domain contracts
(host-tool, host-uri, workflow-gate/schema, unattended types) from
modes/rpc/rpc-types.ts into neutral modes/shared/agent-wire/*-types.ts and
re-point 37 consumers (src + tests). Transport-command types (RpcCommand/
RpcResponse/etc.) stay in rpc-types.ts pending later phases. acp + notifications
untouched.

Gates: bun run check exit 0; focused workflow-gate/unattended suites 148 pass;
zero-import scan clean (no kept file imports moved types from modes/rpc/rpc-types).
