+++ packages/coding-agent/test/harness-control-plane/unattended-lifecycle-redteam.test.ts
@@ -5,7 +5,6 @@ import { describe, expect, it } from "bun:test";
 import { mkdtempSync } from "node:fs";
 import { tmpdir } from "node:os";
 import * as path from "node:path";
-import type { RpcUnattendedDeclaration, RpcWorkflowGate } from "@gajae-code/coding-agent/modes/rpc/rpc-types";
 import {
 	approvalGate,
 	decodeApproval,
@@ -16,6 +15,10 @@ import {
 	gateAnswerToResult,
 	questionToGate,
 } from "@gajae-code/coding-agent/modes/shared/agent-wire/deep-interview-gate";
+import type {
+	RpcUnattendedDeclaration,
+	RpcWorkflowGate,
+} from "@gajae-code/coding-agent/modes/shared/agent-wire/protocol";
 import { UnattendedAuditLog } from "@gajae-code/coding-agent/modes/shared/agent-wire/unattended-audit";
 import {
 	ActionDeniedError,
