+++ packages/coding-agent/test/rpc-host-uris.test.ts
@@ -2,7 +2,10 @@ import { afterEach, describe, expect, it } from "bun:test";
 import { InternalUrlRouter } from "@gajae-code/coding-agent/internal-urls";
 import { parseInternalUrl } from "@gajae-code/coding-agent/internal-urls/parse";
 import { RpcHostUriBridge } from "@gajae-code/coding-agent/modes/rpc/host-uris";
-import type { RpcHostUriCancelRequest, RpcHostUriRequest } from "@gajae-code/coding-agent/modes/rpc/rpc-types";
+import type {
+	RpcHostUriCancelRequest,
+	RpcHostUriRequest,
+} from "@gajae-code/coding-agent/modes/shared/agent-wire/protocol";
 
 const router = InternalUrlRouter.instance();
