+++ packages/coding-agent/test/unattended-action-redteam.test.ts
@@ -1,5 +1,8 @@
 import { describe, expect, it } from "bun:test";
-import type { RpcUnattendedActionClass, RpcUnattendedDeclaration } from "@gajae-code/coding-agent/modes/rpc/rpc-types";
+import type {
+	RpcUnattendedActionClass,
+	RpcUnattendedDeclaration,
+} from "@gajae-code/coding-agent/modes/shared/agent-wire/protocol";
 import { classifyBashAction } from "@gajae-code/coding-agent/modes/shared/agent-wire/unattended-action-policy";
 import {
 	ActionDeniedError,
