+++ packages/coding-agent/test/unattended-session.test.ts
@@ -1,7 +1,10 @@
 import { describe, expect, it } from "bun:test";
-import type { RpcUnattendedDeclaration, RpcWorkflowGate } from "@gajae-code/coding-agent/modes/rpc/rpc-types";
 import { approvalGate } from "@gajae-code/coding-agent/modes/shared/agent-wire/approval-gate";
 import { questionToGate } from "@gajae-code/coding-agent/modes/shared/agent-wire/deep-interview-gate";
+import type {
+	RpcUnattendedDeclaration,
+	RpcWorkflowGate,
+} from "@gajae-code/coding-agent/modes/shared/agent-wire/protocol";
 import {
 	modelSupportsTokenCostMetrics,
 	UnattendedSessionControlPlane,
